Favicon logic and utility to class
